refactor: move the toast copy out of the component, return a record from publish
Both from review.

The message map and its DAO branch are pure data and a pure function, so they no
longer live in the component module. The test now imports a file whose only
import is a type — previously it pulled framer-motion, react-remove-scroll and
the design system in to assert a string.

The publish effect returns `{ spaceGovernanceType }` rather than the bare value.
Nothing is discarded today — `makeProposal` returns void — but a record says at
the call site what the value is for, and lets a future publish result join it
instead of replacing it.
